requirements.txt
flask==2.3.3
twilio==7.17.0
google-generativeai==0.3.0
python-dotenv==0.19.0
flask-caching==1.10.1
gunicorn==20.1.0
flask-talisman==0.8.1
werkzeug==2.3.7 